Update to the current master from Ravenbrook.

+++ code/sacss.c
@@ -1,6 +1,6 @@
 /* sacss.c: SAC MANUAL ALLOC STRESS TEST
  *
- * $Id: //info.ravenbrook.com/project/mps/master/code/sacss.c#13 $
+ * $Id: //info.ravenbrook.com/project/mps/master/code/sacss.c#14 $
  * Copyright (c) 2001 Ravenbrook Limited.  See end of file for license.
  * Portions copyright (C) 2002 Global Graphics Software.
  */
@@ -80,7 +80,7 @@ static mps_res_t stress(mps_class_t class,
   for (k = 0; k < testLOOPS; ++k) {
     /* shuffle all the objects */
     for (i=0; i<testSetSIZE; ++i) {
-      int j = rnd()%(testSetSIZE-i);
+      int j = (int)(rnd()%(unsigned)(testSetSIZE-i));
       void *tp;
       size_t ts;
